Rename .py file
